短信平台

+++ platform-sms-admin/admin-web/src/main/java/com/courage/platform/sms/admin/dao/TSmsAppinfoDAO.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Map;
 
 @Repository
-public interface TSmsAppinfoDAO extends MyBatisBaseDao<TSmsAppinfo, Long> {
+public interface TSmsAppinfoDAO extends MyBatisBaseDao<TSmsAppinfo, Integer> {
 
     TSmsAppinfo getAppinfoByAppKey(@Param("appKey") String appKey);
